user levels added, database doesnt allow empty string entries

+++ pnve3.py
@@ -12,8 +12,12 @@ rel_path = os.path.join(cwd, 'tmp/pnve3.db')
 DATABASE_PATH = rel_path
 DEBUG = True
 SECRET_KEY = 'dev_key'
-USERNAME = 'admin'
-PASSWORD = 'jelszo'
+USERNAME_USER_LEVEL = 'guest'
+PASSWORD_USER_LEVEL = 'userjelszo'
+USERNAME_ENGINEER_LEVEL = 'admin'
+PASSWORD_ENGINEER_LEVEL = 'adminjelszo'
+USERNAME_ADMIN_LEVEL = 'sadmin'
+PASSWORD_ADMIN_LEVEL = 'superadmin'
 
 app = Flask(__name__)
 Bootstrap(app)
@@ -40,6 +44,8 @@ def teardown_request_fromdb(exception):
 
 @app.route('/')
 def show_tbl1xx_entries():
+    if (not session.get('logged_in_guest') and not session.get('logged_in_engineer') and not session.get('logged_in_admin')):
+        return redirect('login')
     template_data = {}
     #cursor_partnumber = g.db.execute('select grp || '-' || substr('00000'||pn,-5,5) || '-' || ver from tbl1xx as partnumber')
     cursor = g.db.execute('SELECT * FROM tbl1xx')
@@ -53,8 +59,8 @@ def show_tbl1xx_entries():
 
 @app.route('/addtbl1xx', methods=['POST'])
 def add_tbl1xx_entry():
-#    if not session.get('logged_in'):
-#        abort(401)
+    if (not session.get('logged_in_admin') and not session.get('logged_in_engineer')):
+        abort(401)
     g.db.execute('INSERT INTO tbl1xx (grp, ver, value, param, desc, status, rohs, datasheet) \
         VALUES (?, ?, ?, ?, ?, ?, ?, ?)',
         [request.form['grp'], request.form['ver'], request.form['value'], request.form['param'], request.form['desc'],
@@ -65,17 +71,15 @@ def add_tbl1xx_entry():
 
 @app.route('/deltbl1xx', methods=['POST'])
 def del_tbl1xx_entry():
-#   if not session.get('logged_in'):
-#       abort(401)
+    if (not session.get('logged_in_admin') and not session.get('logged_in_engineer')):
+       abort(401)
     g.db.execute('DELETE FROM tbl1xx WHERE pn=?',[request.form['delete']])
     g.db.commit()
     flash('Entry was successfully deleted')
     return redirect(url_for('show_tbl1xx_entries'))
 
-@app.route('/mod_tbl1xx_entry_page')
-def mod_tbl1xx_entry_page():
-#    if not session.get('logged_in'):
-#        abort(401)
+@app.route('/mod_tbl1xx_entry_page/<pk>')
+def mod_tbl1xx_entry_page(pk):
     template_data = {}
 #   cursor_partnumber = g.db.execute('select grp || '-' || substr('00000'||pn,-5,5) || '-' || ver from tbl1xx as partnumber')
     cursor = g.db.execute('SELECT * FROM tbl1xx')
@@ -85,17 +89,51 @@ def mod_tbl1xx_entry_page():
     template_data["tbls"] = map(lambda x: x[0], tbl_select.fetchall())
     template_data["cols"] = list(map(lambda x: x[0], cursor.description))
 #   template_data["pn"] = cursor_partnumber_fetchall()
+    template_data["pn"] = pk
     return render_template('mod_form.html', **template_data)
 
-@app.route('/mod_tbl1xx_entry')
+@app.route('/mod_tbl1xx_entry', methods=['POST'])
 def mod_tbl1xx_entry():
-    g.db.execute('UPDATE tbl1xx SET grp = ?, ver = ?, value = ?, param = ?, desc = ?, status = ?, rohs = ?, datasheet = ?) \
+    if not session.get('logged_in_admin'):
+       abort(401)
+    g.db.execute('UPDATE tbl1xx SET grp = ?, ver = ?, value = ?, param = ?, desc = ?, status = ?, rohs = ?, datasheet = ? \
         WHERE pn = ?',
         [request.form['grp'], request.form['ver'], request.form['value'], request.form['param'], request.form['desc'],
-        request.form['status'], request.form['rohs'], request.form['datasheet'], request.form['pn']])
+        request.form['status'], request.form['rohs'], request.form['datasheet'], request.form['modify']])
     g.db.commit()
     flash('New entry was successfully modified')
-    return redirect(url_for('mod_tbl1xx_entry_page'))
+    return redirect(url_for('show_tbl1xx_entries'))
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    permission_error = None
+    if request.method == 'POST':
+        if (request.form['username'] == app.config['USERNAME_USER_LEVEL'] and
+            request.form['password'] == app.config['PASSWORD_USER_LEVEL']):
+                session['logged_in_guest'] = True
+                flash('You were logged in as a guest level user')
+                return redirect(url_for('show_tbl1xx_entries'))
+        elif (request.form['username'] == app.config['USERNAME_ENGINEER_LEVEL'] and
+            request.form['password'] == app.config['PASSWORD_ENGINEER_LEVEL']):
+                session['logged_in_engineer'] = True
+                flash('You were logged in as an engineer level user')
+                return redirect(url_for('show_tbl1xx_entries'))
+        elif (request.form['username'] == app.config['USERNAME_ADMIN_LEVEL'] and
+            request.form['password'] == app.config['PASSWORD_ADMIN_LEVEL']):
+                session['logged_in_admin'] = True
+                flash('You were logged in as an admin level user')
+                return redirect(url_for('show_tbl1xx_entries'))
+        else:
+            permission_error = 'Invalid username or password'
+    return render_template('login.html', permission_error = permission_error)
+
+@app.route('/logout')
+def logout():
+    session.pop('logged_in_guest', None)
+    session.pop('logged_in_engineer', None)
+    session.pop('logged_in_admin', None)
+    flash('You were logged out')
+    return redirect(url_for('login'))
 
 if __name__ == '__main__':
     app.run()
